refactor: move first anonymous flow to a separate file
